(#9): Restructure files for categorisation
